Move IListNode interface to the correct namespace
