+++ SAGESharp.Tests/IO/Binary/TreeBasedSerialization/Trees/Builders.cs
@@ -5,6 +5,7 @@
  */
 using NSubstitute;
 using SAGESharp.IO;
+using SAGESharp.IO.Binary.TreeBasedSerialization;
 using System;
 using System.Collections.Generic;
