+++ SAGESharp/SAGESharp.csproj
@@ -98,6 +98,7 @@
     <Compile Include="IO\Binary\RightPaddingAttribute.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\AbstractOffsetNode.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\Edge.cs" />
+    <Compile Include="IO\Binary\TreeBasedSerialization\IListNode.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\ListNode.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\OffsetNode.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\PaddingNode.cs" />
